+++ library/src/androidMain/kotlin/com/lagradost/cloudstream3/utils/Coroutines.android.kt
@@ -1,0 +1,11 @@
+package com.lagradost.cloudstream3.utils
+
+import android.os.Handler
+import android.os.Looper
+
+actual fun runOnMainThreadNative(work: () -> Unit) {
+    val mainHandler = Handler(Looper.getMainLooper())
+    mainHandler.post {
+        work()
+    }
+}
