+++ library/src/jvmMain/kotlin/com/lagradost/cloudstream3/utils/Coroutines.jvm.kt
@@ -1,0 +1,5 @@
+package com.lagradost.cloudstream3.utils
+
+actual fun runOnMainThreadNative(work: () -> Unit) {
+    work.invoke()
+}
